Fix updating problem
This appears to fix the problem of the applet not refreshing. It only refreshed once because applet_check_uptime was not returning a value. The documentation I found for the function g_timeout_add said the function specified will only be run at regular intervals as long as it don't return FALSE.

https://developer.gnome.org/glib/stable/glib-The-Main-Event-Loop.html#g-timeout-add